backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pandas==2.1.4
python-dotenv==1.0.0
pydantic==2.5.3
numpy==1.26.3 